requirements.txt
aniso8601==8.0.0
attrs==20.3.0
bson==0.5.10
CacheControl==0.12.6
cachelib==0.1.1
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
cryptography==3.3.1
DateTime==4.3
dnspython==1.16.0
ecdsa==0.14.1
firebase-admin==4.5.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
httpie==2.3.0
httplib2==0.18.1
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
msgpack==1.0.2
openapi-spec-validator==0.2.9
passlib==1.7.4
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pymongo==3.11.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
stripe==2.55.0
uritemplate==3.0.1
urllib3==1.25.11
uuid==1.30
Werkzeug==1.0.1
zope.interface==5.2.0
